feat: bundle the Rust fastled CLI binary in the wheel (closes #67)
End users who `pip install fastled` will now receive the Rust `fastled`
binary directly in their venv's Scripts/bin directory — no Python
entry-point shim, no cargo build at install time.

Mechanism: maturin's `[tool.maturin] data` config points at a
`wheel-data/` directory which it packages as the wheel's
`<pkg>-<ver>.data/` tree. Pip extracts `data/scripts/*` to the venv's
Scripts/ (Windows) or bin/ (Unix) on install. Verified locally:

    $ ./build-wheel
    $ unzip -l target/wheels/fastled-2.0.7-cp310-abi3-win_amd64.whl | grep .data
        6303744  fastled-2.0.7.data/scripts/fastled.exe

    $ uv pip install target/wheels/fastled-*.whl
    $ which fastled
    /tmp/test-venv/Scripts/fastled.exe        # the Rust binary
    $ fastled --version
    fastled 2.0.7

Changes:

- `crates/fastled-cli/src/lib.rs` (new) — moved the body of main.rs and
  renamed `fn main` to `pub fn run`. `crates/fastled-cli/src/main.rs`
  shrinks to a one-liner calling `fastled_cli::run()`. The crate now
  exposes both `[lib]` and `[[bin]]`, so the binary's logic can also be
  called from other consumers in the future.
- `pyproject.toml` — adds `[tool.maturin] data = "wheel-data"`; drops
  `fastled = "fastled.cli:main"` from `[project.scripts]` (the Rust
  binary takes over that command). `fled` and `fastled-wasm` stay as
  Python aliases.
- `build-wheel` (new) — helper that compiles the bin via the `_cargo`
  trampoline, stages it under `wheel-data/scripts/`, then runs
  `maturin build --release`.
- `.github/workflows/_build.yml` — adds a pre-step that compiles the bin
  for the target triple and stages it under `wheel-data/scripts/` before
  the existing `maturin build` step. Cross-compile path handled.
- `src/fastled/_rust_cli.py` — drops the `_looks_like_python_dir` PATH
  filter; the Python `fastled` shim no longer exists, so plain
  `shutil.which("fastled")` is sufficient for the wheel-install case.
- `.gitignore` — ignore `wheel-data/`.

Editable installs (CI unit-test, deploy, local dev) are unchanged: the
existing cargo-build fallback in `./install` (PR #66) keeps populating
`target/release/fastled` for the workspace-target lookup. A follow-up
can switch CI's editable installs to wheel installs if desired, but
that's orthogonal to this PR.

Local: `bash lint` clean, 119 unit tests pass.
